Remove SwiftDirectoryWatcher dependency and related files
