+++ Screenshots/Classes/Screenshot.swift
@@ -1,9 +1,4 @@
-
 import Cocoa
-import SwiftDirectoryWatcher
-
-let DEFAULT_MAX_RETRIES: Int = 10
-let DEFAULT_RETRY_WAIT: Double = 0.1
 
 public enum ScreenshotError: String, Error {
   case invalidImage
@@ -40,29 +35,6 @@ public protocol ScreenshotWatcherDelegate {
 
 protocol ScreenshotURL {
   var screenCaptureRect: CGRect? { get }
-}
-
-func readUntilDefined<T>(
-  retries: Int, retryAcc: Int = 0, wait: Double,
-  read: @escaping () -> T?,
-  callback: @escaping (_ result: T?, _ retries: Int) -> Void
-  ) {
-  if let result = read() {
-    callback(result, retryAcc)
-    return
-  }
-  
-  if retries == 0 {
-    callback(nil, retryAcc)
-    return
-  }
-  
-  DispatchQueue.main.asyncAfter(deadline: .now() + wait) {
-    readUntilDefined(
-      retries: retries - 1, retryAcc: retryAcc + 1,
-      wait: wait, read: read, callback: callback
-    )
-  }
 }
 
 func getAttributes(for url: URL) -> CGRect? {
@@ -114,82 +86,7 @@ func getAttributes(for url: URL) -> CGRect? {
  
 }
 
-
-extension URL: ScreenshotURL {
-  var screenCaptureRect: CGRect? {
-    return getAttributes(for: self)
-//    guard let item = NSMetadataItem(url: self) else {
-//      return nil
-//    }
-//
-//    guard let attributes = item.value(forAttribute: "kMDItemScreenCaptureGlobalRect") as? [Int] else {
-//      return nil
-//    }
-//
-//    return CGRect(x: attributes[0], y: attributes[1], width: attributes[2], height: attributes[3])
-  }
-  
-  func readScreenCaptureRect(retries: Int, wait: Double, result: @escaping (CGRect?, Int) -> Void) {
-    readUntilDefined(retries: retries, wait: wait, read: { () -> CGRect? in
-      return self.screenCaptureRect
-    }) { (rect: CGRect?, retries: Int) in
-      result(rect, retries)
-    }
-  }
-}
-
-public class SystemScreenshotWatcher: ScreenshotWatcher {
-  public var maxRetries = DEFAULT_MAX_RETRIES
-  public var retryWait = DEFAULT_RETRY_WAIT
-  
-  public static var shared = SystemScreenshotWatcher()
-  
-  public var delegate: ScreenshotWatcherDelegate?
-  
-  lazy var screenshotDirectory = FileManager.default.urls(for: .desktopDirectory, in: .userDomainMask).first!
-  
-  lazy var directoryWatcher: DirectoryWatcher = {
-    let watcher = DirectoryWatcher(url: screenshotDirectory)
-    watcher.delegate = self
-    return watcher
-  }()
-  
-  public init() {
-    
-  }
-  
-  public func start() {
-    directoryWatcher.start()
-  }
-  
-  public func stop() {
-    directoryWatcher.stop()
-  }
-}
-
-extension SystemScreenshotWatcher: DirectoryWatcherDelegate {
-  public func directoryWatcher(_ watcher: DirectoryWatcher, changed: DirectoryChangeSet) {
-    guard let screenshotURL = (changed.newFiles.first {
-      $0.pathExtension == "png" && $0.lastPathComponent.contains("Screen Shot")
-    }) else {
-      return
-    }
-    
-    screenshotURL.readScreenCaptureRect(retries: maxRetries, wait: retryWait) { (rect, retries) in
-      var screenshot = Screenshot(url: screenshotURL, rect: rect, error: nil, retries: retries)
-
-      if rect == nil {
-        screenshot.error = ScreenshotError.missingMetadataRectProperty
-      }
-      
-      self.delegate?.screenshotWatcher(self, didCapture: screenshot)
-    }
-  }
-}
-
-public class ScreenshotCLI: ScreenshotWatcher {
-  public var maxRetries = DEFAULT_MAX_RETRIES
-  public var retryWait = DEFAULT_RETRY_WAIT
+public class ScreenshotCLI {
 
   public static var shared = ScreenshotCLI()
   
@@ -197,12 +94,6 @@ public class ScreenshotCLI: ScreenshotWatcher {
   public var taskDelegate: ScreenshotTaskDelegate?
   
   public lazy var screenshotDirectory = FileManager.default.urls(for: .cachesDirectory, in: .userDomainMask).first!
-  
-  lazy var directoryWatcher: DirectoryWatcher = {
-    let watcher = DirectoryWatcher(url: screenshotDirectory)
-    watcher.delegate = self
-    return watcher
-  }()
   
   var task: Process?
     
@@ -215,13 +106,6 @@ public class ScreenshotCLI: ScreenshotWatcher {
     
   }
   
-  public func start() {
-    directoryWatcher.start()
-  }
-  
-  public func stop() {
-    directoryWatcher.stop()
-  }
   
   func createScreenshotURL() -> URL {
     return screenshotDirectory
@@ -271,6 +155,7 @@ public class ScreenshotCLI: ScreenshotWatcher {
         print("Error: task.terminationStatus != 0")
         return
       } else {
+        // FIXME: - Check if url exists otherwise it means that a user cancelled the execution
         let attributes = getAttributes(for: url)
         completion(.success(.init(url: url, rect: attributes, error: nil, retries: 0)))
       }
@@ -321,50 +206,4 @@ public class ScreenshotCLI: ScreenshotWatcher {
         }
       }
   }
-}
-
-extension ScreenshotCLI: DirectoryWatcherDelegate {
-  public func directoryWatcher(_ watcher: DirectoryWatcher, changed: DirectoryChangeSet) {
-    guard let screenshotURL = (changed.newFiles.first {
-      $0.pathExtension == "png" && $0.lastPathComponent.contains("Screen Shot")
-    }) else {
-      return
-    }
-    
-    screenshotURL.readScreenCaptureRect(retries: maxRetries, wait: retryWait) { (rect, retries) in
-      var screenshot = Screenshot(url: screenshotURL, rect: rect, error: nil, retries: retries)
-      
-      if rect == nil {
-        screenshot.error = ScreenshotError.missingMetadataRectProperty
-      }
-      
-      self.delegate?.screenshotWatcher(self, didCapture: screenshot)
-    }
-  }
-}
-
-
-extension String {
-
-    /// Create `Data` from hexadecimal string representation
-    ///
-    /// This creates a `Data` object from hex string. Note, if the string has any spaces or non-hex characters (e.g. starts with '<' and with a '>'), those are ignored and only hex characters are processed.
-    ///
-    /// - returns: Data represented by this hexadecimal string.
-
-    var hexadecimal: Data? {
-        var data = Data(capacity: count / 2)
-
-        let regex = try! NSRegularExpression(pattern: "[0-9a-f]{1,2}", options: .caseInsensitive)
-        regex.enumerateMatches(in: self, range: NSRange(startIndex..., in: self)) { match, _, _ in
-            let byteString = (self as NSString).substring(with: match!.range)
-            let num = UInt8(byteString, radix: 16)!
-            data.append(num)
-        }
-
-        guard data.count > 0 else { return nil }
-
-        return data
-    }
-
 }
